rename entity file
