+++ apps/server/src/modules/h5p-editor/entity/index.ts
@@ -1,3 +1,3 @@
 export * from './h5p-content.entity';
 export * from './library.entity';
-export * from './base-entity-with-timestamp.entity';
+export * from './h5p-editor-tempfile.entity';
