+++ apps/server/src/modules/h5p-editor/service/temporary-file-storage.service.ts
@@ -4,7 +4,7 @@ import { S3ClientAdapter } from '@shared/infra/s3-client';
 import { ReadStream } from 'fs';
 import { Readable } from 'stream';
 import { H5pFileDto } from '../controller/dto/h5p-file.dto';
-import { H5pEditorTempFile } from '../entity/base-entity-with-timestamp.entity';
+import { H5pEditorTempFile } from '../entity/h5p-editor-tempfile.entity';
 import { H5P_CONTENT_S3_CONNECTION } from '../h5p-editor.config';
 import { TemporaryFileRepo } from '../repo/temporary-file.repo';
